Reimplemented karma to use SQLite3
In the process, a large number of changes were made to the frontend to
deal with the asynchronicity of SQLite3's bindings (and the extreme
laziness used in this project for integrity).

The filename is fixed, but a pull to change it to a config shouldn't be
hard. That config should be on the Karma behavior--whole-bot databases
should be in a different namespace.