Updated karma schema
Instead of an encoded subject, the database now stores a (kind, subject)
tuple. All the internal code was updated to make this assumption, and
the public API had only one breaking change (in findOrCreate). Indexes
were added as well, to accelerate searches, on this tuple, and to
enforce the uniqueness on the points table.

If you had a previous database version, you'll have to import the old
values manually. For lack of a script, split the old "subject" on the
first "|" into (kind, subject) and insert that.